# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kimdasa Construction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List all customers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /estimates/{id}:
    get:
      operationId: getEstimate
      tags: [estimates]
      summary: Get an estimate by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
    patch:
      operationId: updateEstimate
      tags: [estimates]
      summary: Update an estimate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstimateRequest"
      responses:
        "200":
          description: Updated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      summary: Delete an estimate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /photos/{id}:
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /ai/photo-estimate:
    post:
      operationId: aiPhotoEstimate
      tags: [ai]
      summary: Generate a preliminary AI estimate from project photos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiPhotoEstimateRequest"
      responses:
        "200":
          description: AI photo estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPhotoEstimateResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
          nullable: true
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateCustomerRequest:
      type: object
      properties:
        leadId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
      required:
        - name
    Estimate:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
          nullable: true
        customerId:
          type: integer
          nullable: true
        serviceType:
          type: string
        measurements:
          type: string
          nullable: true
        difficulty:
          type: string
          nullable: true
        height:
          type: string
          nullable: true
        roofPitch:
          type: string
          nullable: true
        materials:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        laborCost:
          type: string
          nullable: true
        materialCost:
          type: string
          nullable: true
        overhead:
          type: string
          nullable: true
        profitMargin:
          type: string
          nullable: true
        internalCost:
          type: string
          nullable: true
        recommendedPrice:
          type: string
          nullable: true
        clientPrice:
          type: string
          nullable: true
        marginPercent:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: string
          nullable: true
        internalNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - serviceType
        - status
        - createdAt
        - updatedAt
    CreateEstimateRequest:
      type: object
      properties:
        leadId:
          type: integer
        customerId:
          type: integer
        serviceType:
          type: string
        measurements:
          type: string
        difficulty:
          type: string
        height:
          type: string
        roofPitch:
          type: string
        materials:
          type: string
        region:
          type: string
        laborCost:
          type: string
        materialCost:
          type: string
        overhead:
          type: string
        profitMargin:
          type: string
        internalCost:
          type: string
        recommendedPrice:
          type: string
        clientPrice:
          type: string
        marginPercent:
          type: string
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: string
        internalNotes:
          type: string
      required:
        - serviceType
...
    AiPhotoEstimateRequest:
      type: object
      properties:
        images:
          type: array
          description: Up to 3 base64 data URLs of project photos (data:image/jpeg;base64,... or raw base64)
          items:
            type: string
          minItems: 1
          maxItems: 3
        serviceHint:
          type: string
          description: Optional service the visitor selected (roofing, siding, kitchen, bathroom, etc.)
        language:
          type: string
          enum: [en, es, pt]
        notes:
          type: string
          description: Optional free-text context from the visitor
      required:
        - images
    AiPhotoEstimateResponse:
      type: object
      properties:
        serviceDetected:
          type: string
          description: Service type the model detected from the photo
        summary:
          type: string
          description: Short human-readable summary of what is visible and the project condition
        scopeBullets:
          type: array
          items:
            type: string
        lowEstimate:
          type: number
          description: Low end of the preliminary cost range in USD
        highEstimate:
          type: number
          description: High end of the preliminary cost range in USD
        unit:
          type: string
          description: Unit the estimate refers to (job, sq ft, linear ft, room)
        confidence:
          type: string
          enum: [low, medium, high]
        recommendation:
          type: string
          description: Next-step recommendation for the homeowner
        disclaimer:
          type: string
        language:
          type: string
      required:
        - summary
        - lowEstimate
        - highEstimate
        - unit
        - confidence
        - recommendation
        - disclaimer
        - language